split powermeter and standamotor into separate files
